Seperate ZMQStreamWriter and ZMQStreamReader
This is to seperate code for the Spyder environment from code for the
target environment. Next I will put them in different directories.
