+++ spyder_unittest/backend/tests/test_pytestworker.py
@@ -7,14 +7,14 @@
 
 # Standard library imports
 import os
-from unittest.mock import call, create_autospec, MagicMock, Mock
+from unittest.mock import create_autospec, MagicMock, Mock
 
 # Third party imports
 import pytest
 
 # Local imports
 from spyder_unittest.backend.pytestworker import SpyderPlugin, main
-from spyder_unittest.backend.zmqstream import ZmqStreamWriter
+from spyder_unittest.backend.zmqwriter import ZmqStreamWriter
 
 
 class EmptyClass:
